+++ CHANGELOG.md
@@ -70,6 +70,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Problems with recursive imports when the intermediate import didn't
     contain any of the requested elements (i.e. no outputs). (#335)
 - Navigate results: fail when no output to generate. Now you get a warning.
+- Makefile: outputs marked as not run by default were listed in the `all`
+  target.
 
 ## [1.4.0] - 2022-10-12
 ### Added